feat: Add QuillNextImage component for custom image blots in QuillEditor
