+++ packages/quill-next-react/src/quill-next-image.css
@@ -8,7 +8,7 @@
   max-width: 100%;
 }
 
-.qn-image-container image {
+.qn-image-container img {
   width: 100%;
   height: 100%;
 }
